requirements.txt
python-dotenv
python-telegram-bot
urllib3
requests
URLObject
firebase-admin